fix scripts

+++ build/run-test-functional-nonet.sh
@@ -8,7 +8,9 @@ if [ -f  app/logs/test.log ]; then
   rm app/logs/test.log
 fi
 
-sudo rm -rf /dev/shm/cache/test/
+if [ -d /dev/shm/cache/test ]; then
+  sudo rm -rf /dev/shm/cache/test/
+fi
 
 app/console --env=test doctrine:mongodb:schema:drop
 app/console --env=test doctrine:mongodb:fixtures:load
